20211110 맵 알고리즘 수정
가지형 맵으로 간단하게 수정. 세부사항은 팀 노션 참고.

+++ Assets/Scripts/System/levels/RegularLevel.cs
@@ -40,7 +40,7 @@ namespace ArcanaDungeon
             rooms[0].SetPosition(0, 0);
             rooms[0].placed = true;
             int radius = (int)levelsize;
-            double startangle = rand.Next(0, 360 / exitnum);
+            int startangle = rand.Next(0, 360 / exitnum);
             int i = 0;
             foreach (Room d in rooms)
             {
@@ -49,7 +49,7 @@ namespace ArcanaDungeon
 
                 int x = (int)(radius * 7 * Math.Sin((2 * Math.PI * i / exitnum) + startangle));
                 int y = (int)(radius * 7 * Math.Cos((2 * Math.PI * i / exitnum) + startangle));
-                d.angle = (2 * Math.PI * i / exitnum) + startangle;
+                d.angle = (int)(2 * Math.PI * i / exitnum) + startangle;
                 d.SetPosition(x, y);
                 d.placed = true;
                 i++;
